+++ source/utility/encryptor.h
@@ -10,16 +10,15 @@ namespace kl {
     {
     public:
         using key = std::vector<uint8_t>;
-
         key keys[5] = {};
 
         encryptor();
 
         void regenerate_keys();
 
-        void encrypt(void* data, uint64_t size) const;
+        void encrypt(void* data, size_t size) const;
 
-        void decrypt(void* data, uint64_t size) const;
+        void decrypt(void* data, size_t size) const;
 
         template <typename T>
         void encrypt(T& obj) const
